agents/agentic_loop_tool_call_gemini_python/pyproject.toml
[project]
name = "agentic-loop-tool-call-gemini-python"
version = "0.1"
description = "Agentic loop with tool calling using Google Gemini"
authors = [{ name = "Temporal Technologies Inc", email = "sdk@temporal.io" }]
requires-python = ">=3.10"
readme = "README.md"
license = "MIT"
dependencies = [
    "temporalio>=1.15.0,<2",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
